fix: add nameserver when resolv is empty on target image

+++ src/mvmctl/models/vm_config_file.py
@@ -37,7 +37,7 @@ def _omit_none(obj: Any) -> Any:
 class VMExportComputeConfig:
     """Compute resources (vcpus, memory)."""
 
-    vcpus: int | None = None
+    vcpus: float | None = None
     mem: int | None = None
